n-parity checker multilayer network
This was the first challenging task across the project. It required implementing the error backpropagation algorithm across multiple layers. I added some improvements later on, like momentum to prevent zigzagging and flattening of the weights and mini-batching, which trades off stability for efficiency of the network (pretty good moves are better than a single perfect one). I definitely need some structure at this point, so I made a layer class to abstract things, which helped a lot.